requirements.txt
requests
slack_sdk
pandas
numpy
google-genai